requirements.txt
pandas>=1.3.0
yfinance>=0.2.0
requests>=2.25.0
pyyaml>=5.4.0
numpy>=1.21.0
lxml>=4.6.0
html5lib>=1.1
